Refactor files in the project structure
