+++ app/src/main/AndroidManifest.xml
@@ -15,14 +15,14 @@
             android:configChanges="orientation|keyboardHidden|locale"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".LanguageActivity"
+            android:name=".activity.LanguageActivity"
             android:configChanges="orientation|keyboardHidden|locale"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".SettingsActivity"
+            android:name=".activity.SettingsActivity"
             android:screenOrientation="portrait" />
         <activity
-            android:name=".AboutActivity"
+            android:name=".activity.AboutActivity"
             android:screenOrientation="portrait" />
