+++ app/src/main/java/com/laine/mauro/currencyrecognition/activity/AboutActivity.kt
@@ -1,7 +1,8 @@
-package com.laine.mauro.currencyrecognition
+package com.laine.mauro.currencyrecognition.activity
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.laine.mauro.currencyrecognition.R
 
 class AboutActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
