+++ app/src/main/java/com/laine/mauro/currencyrecognition/activity/LanguageActivity.kt
@@ -1,9 +1,11 @@
-package com.laine.mauro.currencyrecognition
+package com.laine.mauro.currencyrecognition.activity
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
+import com.laine.mauro.currencyrecognition.activity.adapter.LanguageAdapter
+import com.laine.mauro.currencyrecognition.R
 import kotlinx.android.synthetic.main.activity_language.*
 
 class LanguageActivity : AppCompatActivity() {
