+++ app/src/main/java/com/laine/mauro/currencyrecognition/activity/SettingsActivity.kt
@@ -1,8 +1,9 @@
-package com.laine.mauro.currencyrecognition
+package com.laine.mauro.currencyrecognition.activity
 
 import android.content.Intent
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.laine.mauro.currencyrecognition.R
 import kotlinx.android.synthetic.main.rv_language_row.*
 
 class SettingsActivity : AppCompatActivity() {
